move codebase to /src dir
